requirements.txt
aiohttp==3.8.1
aiohttp-cors==0.7.0
aioredis==1.3.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
blessed==1.19.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorful==0.5.4
Deprecated==1.2.13
fastapi==0.71.0
filelock==3.4.2
frozenlist==1.2.0
google-api-core==2.3.2
google-auth==2.3.3
googleapis-common-protos==1.54.0
gpustat==1.0.0b1
grpcio==1.43.0
h11==0.12.0
hiredis==2.0.0
idna==3.3
importlib-resources==5.4.0
jsonschema==4.3.3
msgpack==1.0.3
multidict==5.2.0
numpy==1.22.0
nvidia-ml-py3==7.352.0
opencensus==0.8.0
opencensus-context==0.1.2
packaging==21.3
Pillow==9.0.0
prometheus-client==0.12.0
protobuf==3.19.3
psutil==5.9.0
py-spy==0.3.11
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-multipart==0.0.5
PyYAML==6.0
ray==1.9.1
redis==4.1.0
requests==2.27.1
rsa==4.8
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
starlette==0.17.1
torch==1.10.1
torchvision==0.11.2
typing-extensions==4.0.1
urllib3==1.26.8
uvicorn==0.16.0
wcwidth==0.2.5
wrapt==1.13.3
yarl==1.7.2
zipp==3.7.0
